[refactor] 입력값 유효성 검증 기능 클래스 분리
