+++ src/main/java/racingcar/application/CarConstants.java
@@ -1,0 +1,5 @@
+package racingcar.application;
+
+public class CarConstants {
+    public static final String CAR_NAME_SPLITTER = ",";
+}
